BIG MOMA SYMBOL CHART
Word is using unicode so i took as many symbols as i can find there are
over 8000 symbols in this chart. it looks messy but its just the same
patern over and over again,i chose this so that it'll be easier do deal
with each symbol if necesery.
* The size of the chart is 44 in a row witch is what word has on my
computer in full screen. it changes as you drag the screen smaller with
your mouse.